Add LiveAccontVerify component and update Test component
