+++ src/pages/test.tsx
@@ -1,26 +1,15 @@
-import fetchNextApi, { apiParamsType } from "@/common/helpers/fetchNextApi";
-import { setCookie } from 'cookies-next';
-import { nextRoutes } from "@/constants/apiPaths";
-import { useEffect, useState } from "react";
 import { useRouter } from 'next/router';
+import { useEffect, useState } from "react";
+import { nextRoutes } from "@/constants/apiPaths";
+import fetchNextApi, { apiParamsType } from "@/common/helpers/fetchNextApi";
+import Loading from '@/common/components/Loading/Loading';
 const Test = () => {
   const router = useRouter();
   const [token, setToken] = useState('')
   useEffect(() => {
     const { code } = router.query;
-    console.log('code:', code);
     if (code) {
       handerVerify(code as string);
-      // if (window.opener) {
-      //   window.opener.postMessage(
-      //     { type: 'auth', token: token },
-      //     'https://sun-live.vercel.app'
-      //   );
-      //   window.close();
-      // } else {
-      //   console.error('No window.opener available');
-      // }
-      // window.close();
     }
   }, [router, token]);
 
@@ -32,11 +21,8 @@ const Test = () => {
     };
     try {
       const result = await fetchNextApi(apiParams);
-      console.log('verify', result);
       if (result.statusCode === 200) {
         setToken(result.token);
-        console.log('result.token:', result.token);
-        
         if (window.opener) {
           window.opener.postMessage(
             { type: 'auth', token: result.token },
@@ -52,13 +38,7 @@ const Test = () => {
     }
   };
   return (
-    <button
-      type="button"
-      className=" w-[150px] h-60 bg-primary-green text-white hover:opacity-80"
-      // onClick={hander}
-      >
-      按鈕
-    </button>
+    <Loading/>
   );
 };
